fix: #71 기능 보완

+++ lpick-backend/src/main/java/com/notfound/lpickbackend/support/command/application/service/QuestionCommandService.java
@@ -61,8 +61,10 @@ public class QuestionCommandService {
 
         Answer answer = question.getAnswer();
 
-        answerCommandRepository.delete(answer);
-
+        if(answer != null) {
+            answerCommandRepository.delete(answer);
+        }
+        
         questionCommandRepository.delete(question);
     }
